pyproject.toml
[project]
name = "dream-gen"
version = "0.1.0"
description = "Living AI Dream Window - A continuously morphing desktop aesthetic experience"
authors = [
    {name = "Luxia", email = "mail.luxia@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = {text = "MIT"}
keywords = ["ai", "generative-art", "stable-diffusion", "desktop-widget", "comfyui"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Artistic Software",
    "Topic :: Multimedia :: Graphics",
]

dependencies = [
    # Core ML dependencies (Note: Install PyTorch with CUDA separately)
    # Use: uv pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "torchaudio>=2.1.0",
    "numpy>=1.24.0,<2.0.0",
    "pillow>=10.1.0",
    # AI/ML
    "diffusers>=0.25.0",
    "huggingface-hub>=0.19.0",
    # API communication
    "requests>=2.31.0",
    "websockets>=12.0",
    "aiohttp>=3.9.0",
    # Configuration
    "pyyaml>=6.0.1",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    # Utilities
    "python-dateutil>=2.8.2",
    "tqdm>=4.66.0",
    "rich>=13.7.0",
    "psutil",
    # Video generation
    "imageio>=2.33.0",
    "imageio-ffmpeg>=0.4.9",
    # Tools for similarity analysis and metrics comparison
    "imagehash>=4.3.1",          # Required for pHash-8 (production encoder)
    "scikit-image>=0.22.0", 
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    
    # Code quality
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
    
    # Type stubs
    "types-pyyaml>=6.0.12",
    "types-requests>=2.31.0",
    "types-pillow>=10.1.0",
]

windows = [
    # Windows-specific dependencies
    "pywin32>=306; platform_system=='Windows'",
    "psutil>=5.9.6; platform_system=='Windows'",
]

all = [
    "dream-gen[dev,windows]",
]

[project.urls]
Homepage = "https://github.com/luxia/dream-gen"
Documentation = "https://github.com/luxia/dream-gen/tree/main/docs"
Repository = "https://github.com/luxia/dream-gen"
Issues = "https://github.com/luxia/dream-gen/issues"

[project.scripts]
dream-gen = "backend.main:main"
dream-controller = "backend.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.black]
line-length = 100
target-version = ['py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_calls = true
disallow_any_generics = false
ignore_missing_imports = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--cov=backend",
    "--cov-report=term-missing",
    "--cov-report=html",
]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

# UV configuration for PyTorch CUDA packages
[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
# PyTorch packages should be installed from the CUDA index
torch = { index = "pytorch-cu121" }
torchvision = { index = "pytorch-cu121" }
torchaudio = { index = "pytorch-cu121" }

